F19 : display advert information except for images

+++ app/src/main/java/c0d3/vitreen/app/fragments/adding/Adding2Fragment.kt
@@ -140,7 +140,7 @@ class Adding2Fragment : ChildFragment() {
                                                     Advert(
                                                             title = title,
                                                             description = description,
-                                                            price = price.toFloat(),
+                                                            price = price.toLong(),
                                                             brand = editTextBrand.text.toString(),
                                                             size = editTextSize.text.toString(),
                                                             locationId = locationId,
