+++ app/src/main/java/c0d3/vitreen/app/models/Advert.kt
@@ -3,10 +3,10 @@ package c0d3.vitreen.app.models
 data class Advert(
     val title: String,
     val description: String,
-    val price: Float,
+    val price: Long,
     val brand: String,
     val size: String? = null,
-    val numberOfConsultations: Int = 0,
+    val numberOfConsultations: Long = 0,
     val reported: ArrayList<String>? = null,
     val locationId: String,
     val categoryId: String,
